debug

+++ app/DTOs/Payment/PaymentRequestData.php
@@ -46,11 +46,12 @@ class PaymentRequestData
 
         if (setting('tax') > 0) {
             $extraItems[] = [
-                'name' => Str::limit( 'Tax', 50, ''),
-                'amount' => (int) round(((int) getPriceAfterTax(setting('tax'), $this->order->subtotal)) * 100),
+                'name' => Str::limit('Tax', 50, ''),
+                'amount' => (int) round(getPriceAfterTax(setting('tax'), $this->order->subtotal) * 100),
                 'quantity' => 1,
             ];
         }
+
         if ($this->order?->discount_amount > 0) {
             $extraItems[] = [
                 'name' => Str::limit('Discount', 50, ''),
